requirements.txt
nose == 1.3.3
mock == 2.0.0
pyhamcrest == 1.8.0
coverage == 3.7.1
radon == 1.2.1
pylint == 1.6.5
pep8 == 1.5.7
pep257 == 0.3.2
flake8 == 2.4.1
click == 3.3
PyYAML == 3.11
Jinja2 == 2.8
markdown == 2.6.2
nose-watch == 0.9.1
